Flask_based_Text_Similarity/requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
MarkupSafe==1.1.1
mysql-connector-python-rf==2.2.2
nltk==3.4.4
numpy==1.17.0
pkg-resources==0.0.0
scikit-learn==0.21.3
scipy==1.3.0
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
Werkzeug==0.15.5
